requirements.txt
Flask==2.3.2
python-docx==0.8.11
openai==0.27.8
werkzeug==2.3.6
python-dotenv==1.0.0